Fixed timer bug

+++ AVR System/lib/LSTD_TYPES.h
@@ -2,13 +2,13 @@
 #define LSTD_TYPES_H
 typedef unsigned char 		u8;
 typedef unsigned short 		u16;
-typedef unsigned int 		u32;
-typedef unsigned long int 	u64;
+typedef unsigned long int 	u32;
+typedef unsigned long long 	u64;
 
-typedef char 				s8;
-typedef short 				s16;
-typedef int 				s32;
-typedef long int 			s64;
+typedef signed char 				s8;
+typedef signed short 				s16;
+typedef signed long int 			s32;
+typedef signed long long 			s64;
 
 typedef unsigned int		uint;
 typedef int 				sint;
